requirements.txt
# Leitura de Documentos
python-docx
pypdf
numpy
pandas
matplotlib

# Vector Store
faiss-cpu

# Bibliotecas de IA
openai
anthropic

# Utilidades
python-dotenv